#2033 renamed VenueFormater to VenueFormatter and moved it to package utils.formatter
