+++ selenium/qa/src/test/java/test/facade/orders/order/OrderManageFacade.java
@@ -21,15 +21,13 @@ import pages.components.admin.orders.manage.ActivityItem.ExpandedContent;
 import pages.components.admin.orders.manage.ActivityItem.NoteExpandedContent;
 import pages.components.admin.orders.manage.ActivityItem.RefundedExpandedContent;
 import pages.components.admin.orders.manage.OrderInfo;
-import pages.components.admin.orders.manage.tickets.TicketRow;
 import pages.components.admin.orders.manage.tickets.OrderTicketsDetails.PerOrderFeeComponent;
+import pages.components.admin.orders.manage.tickets.TicketRow;
 import pages.components.dialogs.IssueRefundDialog;
 import pages.components.dialogs.IssueRefundDialog.RefundReason;
 import test.facade.BaseFacadeSteps;
-import test.facade.FacadeProvider;
 import utils.MsgConstants;
 import utils.ProjectUtils;
-import utils.VenueFormater;
 
 public class OrderManageFacade extends BaseFacadeSteps{
